junit: Implement fuzzing within JUnit Jupiter
Instead of implementing fuzzing runs as a custom JUnit test engine, this
commit implements it within the JUnit Jupiter platform instead.

The main benefit of this approach is the full and effortless
interoperability with lifecycle hooks such as `@BeforeAll` and
`@AfterAll`. The new implementation is also significantly shorter and
automatically avoids inconsistencies such as
https://github.com/CodeIntelligenceTesting/jazzer/issues/539 and
https://github.com/CodeIntelligenceTesting/jazzer/issues/543.

The new implementation models fuzzing as a single invocation of the
`@ParameterizedTest` used for the regression test with fake arguments
named `Fuzzing...`. The registered `InvocationInterceptor` is then used
to replace the invocation with the actual fuzzing run.

In this way, `@FuzzTest` methods invoke `@BeforeEach` hooks once before
fuzzing starts. Having the invocation represented as a single invocation
parametrized test should make this behavior less surprising.

Fixes #543